Move createLedgerChannel method into cmd/utils package
