+++ cmd/utils/utils.go
@@ -6,6 +6,13 @@ import (
 	"os/exec"
 	"os/signal"
 	"syscall"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/rs/zerolog"
+	"github.com/statechannels/go-nitro/internal/logging"
+	"github.com/statechannels/go-nitro/internal/testdata"
+	"github.com/statechannels/go-nitro/rpc"
+	"github.com/statechannels/go-nitro/types"
 )
 
 // WaitForKillSignal blocks until we receive a kill or interrupt signal
@@ -29,4 +36,33 @@ func StopCommands(cmds ...*exec.Cmd) {
 			panic(err)
 		}
 	}
+}
+
+func CreateLogger(logFileName string, clientName string) (zerolog.Logger, *os.File) {
+	logDestination := logging.NewLogWriter("./artifacts", logFileName)
+
+	return zerolog.New(logDestination).
+		Level(zerolog.TraceLevel).
+		With().
+		Timestamp().
+		Str("client", clientName).
+		Str("rpc", "client").
+		Logger(), logDestination
+}
+
+func CreateLedgerChannel(client rpc.RpcClientApi, counterPartyAddress common.Address) error {
+	clientAddress, err := client.Address()
+	if err != nil {
+		return err
+	}
+	ledgerChannelDeposit := uint(5_000_000)
+	asset := types.Address{}
+	outcome := testdata.Outcomes.Create(clientAddress, counterPartyAddress, ledgerChannelDeposit, ledgerChannelDeposit, asset)
+	response, err := client.CreateLedgerChannel(counterPartyAddress, 0, outcome)
+	if err != nil {
+		return err
+	}
+
+	<-client.ObjectiveCompleteChan(response.Id)
+	return nil
 }
